implémentation de l'aggégation d'emails
