+++ __main__.py
@@ -1,5 +1,6 @@
-from emails_managment.email_processing import EmailProcessing
-from database_managment.email_database import *
+from hasher.hasher import Hasher
+from aggregator.email_aggregator import EmailAggregator
+from file_retriever.file_retriever import FileRetriever
 
 def paths_to_dict():
     """This is a temporary function for project development..."""
@@ -12,13 +13,9 @@ def paths_to_dict():
                 paths[k] = v.strip()
     return paths
 
-
-
-if __name__ == "__main__":
+if __name__ == '__main__':
     path = paths_to_dict()['emails_eml']
-    ep = EmailProcessing(path=path, with_attachments=True)
-    emails = ep.get_emails_list
-    print(ep.log_execution_time)
-    el = EmailLoader(emails=emails)
-    print(el.log_execution_time)
-    print(el.show_buffer)
+    hasher = Hasher()
+    supported_extensions = ['.OUTLOOK.COM', '.eml', '.mbox']
+    file_retriever = FileRetriever(path=path, supported_extensions=supported_extensions)
+    aggregator = EmailAggregator(hasher, file_retriever)
